refactor(collaborator-avatars): move collaborator-avatars
