+++ i18n/en-US.properties
@@ -440,6 +440,18 @@ boxui.classification.addClassification = Add Classification
 boxui.classification.classification = Classification
 # Button to edit classification on an item
 boxui.classification.edit = Edit
+# Text to display for users who have not accepted an invitation to collaborate yet
+boxui.collaboratorAvatars.collaboration.pendingCollabText = Pending
+# Label for collaborator avatars
+boxui.collaboratorAvatars.collaboratorAvatarsLabel = Shared with
+# Tooltip text for collaborator expiration badge
+boxui.collaboratorAvatars.collaborators.expirationTooltipText = Access expires on {date}
+# This tooltip indicates that a collaborator is not in the same enterprise of the current user
+boxui.collaboratorAvatars.collaborators.externalCollabTooltipText = {email} is from outside of your company
+# Manage all link text on collaborator list
+boxui.collaboratorAvatars.manageAllLinkText = Manage All
+# Text to display in collaborator list if there are more items
+boxui.collaboratorAvatars.viewAdditionalPeopleText = View additional people
 # Icon title for a Box item of type bookmark or web-link
 boxui.contentExplorer.bookmark = Bookmark
 # Aria label for the folder breadcrumb
@@ -1000,20 +1012,12 @@ boxui.shareMenu.viewOnly = View Only
 boxui.unifiedShare.betaFeedbackCtaCopy = Click to provide feedback on this simplified sharing beta experience.
 # Text to display for a group of users who have accepted an invitation to collaborate
 boxui.unifiedShare.collaboration.groupCollabText = Group
-# Text to display for users who have not accepted an invitation to collaborate yet
-boxui.unifiedShare.collaboration.pendingCollabText = Pending
 # Text to display for individual users who have accepted an invitation to collaborate
 boxui.unifiedShare.collaboration.userCollabText = User
-# Label for collaborator avatars
-boxui.unifiedShare.collaboratorAvatarsLabel = Shared with
 # Title for collaborator list modal
 boxui.unifiedShare.collaboratorListTitle = People in '{itemName}'
 # This string is displayed as tooltip on hovering over expire icon for collab
 boxui.unifiedShare.collaborators.expirationTooltipClickableText = Access expires on {date}. Click for details.
-# Tooltip text for collaborator expiration badge
-boxui.unifiedShare.collaborators.expirationTooltipText = Access expires on {date}
-# This tooltip indicates that a collaborator is not in the same enterprise of the current user
-boxui.unifiedShare.collaborators.externalCollabTooltipText = {email} is from outside of your company
 # Text used in button label to describe permission level - co-owner
 boxui.unifiedShare.coownerLevelButtonLabel = Invite as Co-owner
 # Description for Co-owner permission level in permissions table
@@ -1054,8 +1058,6 @@ boxui.unifiedShare.inviteFieldLabel = Invite People
 boxui.unifiedShare.linkShareOff = Link sharing is off
 # Link sharing is on
 boxui.unifiedShare.linkShareOn = Link sharing is on
-# Manage all link text on collaborator list
-boxui.unifiedShare.manageAllLinkText = Manage All
 # Label for "Message" text box to email the Shared Link
 boxui.unifiedShare.message = Message
 # Title of the Unified Share Modal. {itemName} is the name of the file / folder being shared
@@ -1138,8 +1140,6 @@ boxui.unifiedShare.uploaderLevelButtonLabel = Invite as Uploader
 boxui.unifiedShare.uploaderLevelDescription = Upload only
 # Text for Uploader permission level in invitee permission dropdown
 boxui.unifiedShare.uploaderLevelText = Uploader
-# Text to display in collaborator list if there are more items
-boxui.unifiedShare.viewAdditionalPeopleText = View additional people
 # Text used in button label to describe permission level - viewer
 boxui.unifiedShare.viewerLevelButtonLabel = Invite as Viewer
 # Description for Viewer permission level in invitee permission dropdown
