+++ src/features/collaborator-avatars/CollaboratorListItem.js
@@ -3,10 +3,10 @@ import React from 'react';
 import classnames from 'classnames';
 import { FormattedMessage } from 'react-intl';
 
-import { Link } from '../../../components/link';
+import { Link } from '../../components/link';
 
-import { COLLAB_GROUP_TYPE, COLLAB_PENDING_TYPE } from '../constants';
-import messages from '../messages';
+import { COLLAB_GROUP_TYPE, COLLAB_PENDING_TYPE } from './constants';
+import messages from './messages';
 import CollaboratorAvatarItem from './CollaboratorAvatarItem';
 
 type Props = {
