+++ src/features/collaborator-avatars/index.js
@@ -1,0 +1,3 @@
+// @flow
+export { default as CollaboratorAvatars } from './CollaboratorAvatars';
+export { default as CollaboratorList } from './CollaboratorList';
